Update to bevy 6.0 (#6)
* WIP - still can't render player..

* render player again (texture atlas) - add debug logs

* add action

* add OS deps

* update README - verify cache hit in action

+++ src/systems/animate_player.rs
@@ -52,7 +52,7 @@ pub fn animate_player(
         }
 
         if timer.finished() {
-            sprite.index = player.animation_index;
+            sprite.index = player.animation_index as usize;
         }
     }
 }
